backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
cryptography==45.0.7
distro==1.9.0
Django==4.2.25
django-allauth==65.11.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
dotenv==0.9.9
exceptiongroup==1.3.0
fastuuid==0.14.0
filelock==3.19.1
frozenlist==1.8.0
fsspec==2025.10.0
google==3.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.0.1
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jwt==1.4.0
litellm==1.79.1
MarkupSafe==3.0.3
multidict==6.7.0
openai==2.6.1
packaging==25.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==1.7.1
python-dotenv==1.1.1
PyYAML==6.0.3
referencing==0.36.2
regex==2025.10.23
requests==2.32.5
rpds-py==0.27.1
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0
zipp==3.23.0
